# Improve proposals listing performance after cache implementation