--- a/decidim-assemblies/db/migrate/20210310120444_add_followable_counter_cache_to_assemblies.rb
+++ b/decidim-assemblies/db/migrate/20210310120444_add_followable_counter_cache_to_assemblies.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class AddFollowableCounterCacheToAssemblies < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_assemblies, :follows_count, :integer, null: false, default: 0, index: true
+
+    reversible do |dir|
+      dir.up do
+        Decidim::Assembly.reset_column_information
+        Decidim::Assembly.find_each do |record|
+          record.class.reset_counters(record.id, :follows)
+        end
+      end
+    end
+  end
+end