--- a/decidim-core/lib/decidim/followable.rb
+++ b/decidim-core/lib/decidim/followable.rb
@@ -6,7 +6,12 @@ module Decidim
     extend ActiveSupport::Concern
 
     included do
-      has_many :follows, as: :followable, foreign_key: "decidim_followable_id", foreign_type: "decidim_followable_type", class_name: "Decidim::Follow"
+      has_many :follows,
+               as: :followable,
+               foreign_key: "decidim_followable_id",
+               foreign_type: "decidim_followable_type",
+               class_name: "Decidim::Follow",
+               counter_cache: :follows_count
       has_many :followers, through: :follows, source: :user
     end