--- a/decidim-initiatives/db/migrate/20210310120720_add_followable_counter_cache_to_initiatives.rb
+++ b/decidim-initiatives/db/migrate/20210310120720_add_followable_counter_cache_to_initiatives.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class AddFollowableCounterCacheToInitiatives < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_initiatives, :follows_count, :integer, null: false, default: 0, index: true
+
+    reversible do |dir|
+      dir.up do
+        Decidim::Initiative.reset_column_information
+        Decidim::Initiative.find_each do |record|
+          record.class.reset_counters(record.id, :follows)
+        end
+      end
+    end
+  end
+end