--- a/decidim-participatory_processes/db/migrate/20210310120750_add_followable_counter_cache_to_participatory_processes.rb
+++ b/decidim-participatory_processes/db/migrate/20210310120750_add_followable_counter_cache_to_participatory_processes.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class AddFollowableCounterCacheToParticipatoryProcesses < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_participatory_processes, :follows_count, :integer, null: false, default: 0, index: true
+
+    reversible do |dir|
+      dir.up do
+        Decidim::ParticipatoryProcess.reset_column_information
+        Decidim::ParticipatoryProcess.find_each do |record|
+          record.class.reset_counters(record.id, :follows)
+        end
+      end
+    end
+  end
+end