#### :tophat: What? Why?
After caching was added to proposals in #6808, the proposals listing page became very sluggish to load, even when results are being loaded from cache. You can see the effect by browsing Metadecidim's proposals components.

After some investigation, the reason for the poor performance turned out to be the cache hash generation for the proposal cards and particularly the line that takes the followers of the proposal into account here:
https://github.com/decidim/decidim/blob/b7a2ac876cb96a7fb553977f5a864bd9186f0307/decidim-proposals/app/cells/decidim/proposals/proposal_m_cell.rb#L141

Implementing a counter cache for the follows improves the performance greatly as my investigation shows below. These are the average response times with 10 consecutive requests for the proposals listing page with cache disabled and in the development environment:

- Current implementation with poorly performing `cache_hash`: 23,47s / req
- `cache_hash` removed (i.e. no caching for the proposal cards, as it was before #6808): 11,58s / req
- With follows counter cache (after this PR): 11,26s / req

**EDIT NOTE**: These numbers were taken before implementing all the changes, see discussion below. After all the changes I've run the same test with 10 requests averaging at 11,35s / req (I ran it 3 times and it was consistent).

Further discussion is available at:
https://github.com/decidim/decidim/discussions/7574

#### :pushpin: Related Issues
- Related to #6808, #7574

#### Testing
1. Create a Decidim instance with the default seed data
2. Get the URL to one of the proposals listing pages from your instance
3. Run the following command to see the response time of the page for 10 consecutive requests:
```bash
$ for i in {1..10}; do curl -so /dev/null -w '%{time_total}\n' http://localhost:3000/URL_TO_PROPOSALS_COMPONENT; done
```
4. See the performance change when comparing before and after this PR  

#### :clipboard: Checklist
- [x] :question: **CONSIDER** adding a unit test if your PR resolves an issue.
- [x] :heavy_check_mark: **DO** check open PR's to avoid duplicates.
- [x] :heavy_check_mark: **DO** keep pull requests small so they can be easily reviewed.
- [x] :heavy_check_mark: **DO** build locally before pushing.
- [x] :heavy_check_mark: **DO** make sure tests pass.
- [ ] :heavy_check_mark: **DO** make sure any new changes are documented in `docs/`.
- [ ] :heavy_check_mark: **DO** add and modify seeds if necessary.
- [ ] :heavy_check_mark: **DO** add CHANGELOG upgrade notes if required.
- [ ] :heavy_check_mark: **DO** add to GraphQL API if there are new public fields.
- [ ] :heavy_check_mark: **DO** add link to MetaDecidim if it's a new feature.
- [x] :x:**AVOID** breaking the continuous integration build.
- [x] :x:**AVOID** making significant changes to the overall architecture.