@ahukkanen 2021-03-11
I have now fixed the test that broke by including author cache keys in the cache hash instead of the coauthorship cache keys. According to my quick tests, there are no significant differences in the performance between these two approaches when requesting 20 and 100 records per page cached or not cached.

I also applied few more smaller performance improvements in the cache hash generation which might or might not have a small impact. There is still a small performance hit from checking the card image but I think it is needed. It could be also bypassed by updating the proposal cache hash when any of its attachments are updated but it would likely require changing the logic of the attachments handling where other work is being done right now (#7452).

All in all, according to my various tests this improves the performance by 30-50% (in different situations and configurations) compared to the current implementation. So it's still not optimal but it's a step forward. Your mileage may vary, so feel free to also test it yourself and compare with the current implementation (without follows counter cache) to get a better sense of the performance impacts.