@alecslupu 2021-03-11
@ahukkanen , it would be very useful to have in the above key the you would take into consideration the machine translation mechanism. 
```
body: {"ro"=>"Un text aici care trebuie tradus", "en" =>"A simple text that needs to be translated", "machine_translations"=>{"bg"=>"моля, изпратете го", "cs"=>"zašlete to prosím", "da"=>"send venligst dette", "de"=>"bitte senden Sie diese", "el"=>"παρακαλείστε να αποστείλετε το παρόν έγγραφο", "en"=>"please send this", "es"=>"le rogamos envíe el presente", "et"=>"palun saatke see", "fi"=>"lähetä tämä", "fr"=>"veuillez l’envoyer", "ga"=>"seol é seo le do thoil", "hr"=>"molimo vas da pošaljete", "hu"=>"kérjük, küldje el a következő dokumentumot:", "it"=>"si prega di inviare la presente", "lt"=>"prašome siųsti", "lv"=>"lūdzu, nosūtiet šo", "mt"=>"jekk jogħġbok ibgħat dan", "nl"=>"gelieve dit te versturen.", "pl"=>"prosimy o przesłanie", "pt"=>"por favor, envie-o", "sk"=>"zašlite prosím", "sl"=>"prosimo, da to pošljite.", "sv"=>"vänligen skicka det här"}}
```

in the above example, having machine translations enabled, the cache keys you have put in place will always render the `body["en"]` key for `I18n.locale=:en`, and it will  not display `body["machine_translations"]["en"]` if the user will chose to use the functionality.