@ahukkanen 2021-03-11
@alecslupu I have not put any cache keys in place. I am just fixing a performance issue related to the added cache functionality.

What you refer to would be solved by changing the method used to save the machine translations for an object:
https://github.com/decidim/decidim/blob/53b3067ceeeef8de8477c4372cd24cca9858c9a9/decidim-core/app/jobs/decidim/machine_translation_save_job.rb#L28

Changing this to `resource.update(field_name.to_sym => resource[field_name])` would fix the problem you are referring to.

But as said, I am only trying to fix a performance issue related to the cache key generation. Fixing every possible case related to the added cache functionality in #6808 is not in the scope of this PR. There are also other remaining issues related to the added cache functionality such as #7564.

I suggest you create a new bug report for the problem you are describing.