@ahukkanen 2021-03-16
Should we backport this to 0.24 @mrcasals ?

This is a bug that will be shipped with 0.24 if it is not backported...